+++ app/Main.hs
@@ -8,10 +8,12 @@ import System.IO
 import Data.ByteString.Builder
 import Control.Monad
 import Data.Time
+import Data.List
 
-data State = State {time :: Time, sound :: Sound}
-data Time = Time {year :: Int, month :: Int, day :: Int, hour :: Int, minute :: Int}
+data State = State {time :: Time, sound :: Sound, battery :: Battery}
+data Time  = Time {year :: Int, month :: Int, day :: Int, hour :: Int, minute :: Int}
 data Sound = Sound {volume :: Int, muted :: Bool}
+data Battery = Battery {capacity :: Int, charging :: Bool}
 
 main :: IO ()
 main = do
@@ -20,7 +22,8 @@ main = do
   forever $ do
     t <- getTime zone
     s <- getSound
-    hPutBuilder stdout . prettyState $ State t s
+    b <- getBattery
+    hPutBuilder stdout . prettyState $ State t s b
     threadDelay 1000000
 
 getTime :: TimeZone -> IO Time
@@ -36,16 +39,32 @@ getSound = do
   let [active, _db, vol] = take 3 . reverse . words $ out
   pure $ Sound (read . takeWhile (/='%') $ drop 1 vol) (active == "[on]")
 
+getBattery :: IO Battery
+getBattery = do
+  c <- readFile "/sys/class/power_supply/BAT0/capacity"
+  s <- readFile "/sys/class/power_supply/BAT0/status"
+  pure $ Battery (read c) (not $ "Discharging" `isPrefixOf` s)
+
 prettyTime :: Time -> Builder
-prettyTime Time{..} = intDec year <> char7 '-' <> f month <> char7 '-' <> f day <> char7 'T' <> f hour <> char7 ':' <> f minute
-  where f n = let (a,b) = n `divMod ` 10 in intDec a <> intDec b
+prettyTime Time{..} = intDec year <> char7 '.' <> intDec month <> char7 '.' <> intDec day <> char7 ' ' <> intDec hour <> char7 ':' <> padNum '0' 2 minute
+
+padNum :: Char -> Int -> Int -> Builder
+padNum c len i | len > 1 && 10^(len-1) > i = char7 c <> padNum c (len-1) i
+padNum _ _ i = intDec i
 
 prettySound :: Sound -> Builder
-prettySound Sound{..} | muted = "[mute]" <> char7 ' ' <> vol
-                      | otherwise = vol
-  where vol = intDec volume <> char7 '%'
+prettySound Sound{..} = text <> num
+                      where num = padNum ' ' 3 volume <> char7 '%'
+                            text | muted     = " Muted:"
+                                 | otherwise = "Volume:"
+
+prettyBattery :: Battery -> Builder
+prettyBattery Battery{..} = text <> num
+  where num = padNum ' ' 3 capacity <> char7 '%'
+        text | charging  = "Charging:"
+             | otherwise = "Capacity:"
 
 prettyState :: State -> Builder
-prettyState (State{..}) = prettySound sound <> char7 ' ' <> prettyTime time <> char7 '\n'
+prettyState (State{..}) = prettyBattery battery <> "  │  " <>  prettySound sound <> "  │  " <> prettyTime time <> char7 '\n'
 
-  
+