+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for minbar
 
+## master
+
+* Implement Battery, Volume and Time
+
 ## 0.1.0.0 -- 2024-09-02
 
 * First version. Released on an unsuspecting world.